rename merged new gene variant
