+++ runscripts/fireworks/fw_queue.py
@@ -225,7 +225,7 @@ from typing import Any, Dict, List, Optional, Union
 
 from fireworks import FiretaskBase, Firework, LaunchPad, Workflow, ScriptTask
 
-from models.ecoli.sim.variants.new_gene import (NEW_GENE_EXPRESSION_FACTORS,
+from models.ecoli.sim.variants.new_gene_internal_shift import (NEW_GENE_EXPRESSION_FACTORS,
 	NEW_GENE_TRANSLATION_EFFICIENCY_VALUES, NEW_GENE_INDUCTION_GEN,
 	NEW_GENE_KNOCKOUT_GEN)
 from wholecell.fireworks.firetasks import InitRawDataTask
